SQL/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.7.2
asn1==2.7.0
asttokens==2.4.1
autocommand==2.2.2
backports.tarfile==1.2.0
blessed==1.20.0
bpylist2==4.1.1
cached-property==1.5.2
case-insensitive-dictionary==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.38.0
coapthon3==1.0.1
colorama==0.4.6
coloredlogs==15.0.1
construct==2.10.70
cryptography==42.0.2
daemonize==2.5.0
decorator==5.1.1
deprecated==1.2.14
developer-disk-image==0.0.2
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
django-sendgrid-v5==1.2.3
django-sendgrid==1.0.1
django==5.0.1
djangorestframework==3.14.0
dnspython==2.6.1
drf-yasg==1.21.7
editor==1.6.6
email-validator==2.2.0
enum-compat==0.0.3
executing==2.1.0
fastapi-cli==0.0.5
fastapi==0.112.2
gpxpy==1.5.0
h11==0.14.0
hexdump==3.3
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
ifaddr==0.2.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
inflect==7.3.1
inflection==0.5.1
inquirer3==0.6.1
ipsw-parser==1.3.7
ipython==8.27.0
itsdangerous==2.2.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.1
jinja2==3.1.4
jinxed==1.3.0
jwcrypto==1.5.1
la-panic==0.5.0
lzfse==0.4.2
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.3.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
oauthlib==3.2.2
opack==0.1.0
orjson==3.10.7
packaging==23.2
parameter-decorators==0.0.2
parso==0.8.4
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
plumbum==1.8.3
prompt-toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.3
py-air-control==2.3.0
pycparser==2.21
pycrashreport==1.2.4
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic-core==2.20.1
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic==2.8.2
pygments==2.18.0
pygnuutils==0.1.1
pyimg4==0.8.6
pykdebugparser==1.2.4
pylzss==0.3.4
pymobiledevice3==4.13.5
pymysql==1.1.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.9
python-pcapng==2.1.1
pytun-pmd3==2.0.9
pytz==2024.1
pyusb==1.2.1
pywin32==306
pyyaml==6.0.1
qh3==1.0.9
readchar==4.2.0
remotezip==0.12.3
requests==2.31.0
rich==13.8.0
runs==1.2.2
sendgrid==6.11.0
setuptools==74.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
srptools==1.0.1
sslpsk-pmd3==1.0.3
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.38.4
termcolor==2.4.0
tomli==2.0.1
tqdm==4.66.5
traitlets==5.14.3
typeguard==4.3.0
typer==0.12.5
types-python-dateutil==2.9.0.20240821
typing-extensions==4.12.2
tzdata==2023.4
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.30.6
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
xmod==1.8.1
xonsh==0.18.3
zeroconf==0.133.0
zipp==3.19.2
